server/requirements.txt
addict==2.4.0
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
contourpy==1.2.0
cycler==0.12.1
dill==0.3.6
dlib==19.24.1
docker-pycreds==0.4.0
exceptiongroup==1.2.2
face_alignment==1.3.4
fastapi==0.115.5
filelock==3.13.1
fonttools==4.50.0
ftfy==6.2.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.42
h11==0.14.0
idna==3.6
imageio==2.34.0
joblib==1.2.0
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
matplotlib==3.7.1
networkx==3.2.1
numba==0.59.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
packaging==24.0
pandas==2.0.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.25.3
psutil==5.9.8
pydantic==2.10.1
pydantic_core==2.27.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.17
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.43.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
threadpoolctl==3.4.0
tifffile==2024.2.12
torch==1.13.1+cu116
torchmetrics==0.11.4
torchvision==0.14.1+cu116
tqdm==4.66.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.32.1
wandb==0.15.10
wcwidth==0.2.13
